requirements.txt
python-dotenv==1.1.0
langchain==0.3.25
langchain-openai==0.3.23
langchain-core==0.3.65
langgraph==0.4.8
langgraph-prebuilt==0.2.2
langchain-tavily==0.2.6
fastapi==0.115.5
uvicorn==0.32.1